requirements.txt
################################################################################
## Requirements for jd2-server.py
################################################################################

Flask>=3.0.0
myjdapi>=1.1.3
gunicorn>=21.2.0

################################################################################
## Requirements for anime-dl.py
################################################################################

watchdog>=3.0.0
requests>=2.0.0
pyyaml>=5.1
